Rewrite model check as validation check re #10079 (#11646)
* Rewrite model check as validation check re #10079

This model check should have been skippped by checking
the `databases` kwarg provided when `check()` is called,
but that reduces the circumstances on which this check is
run to either:
- manage.py migrate
- manage.py check --databases

So, given how infrequently this would be run, better to just
implement this in our validation framework. This also prevents
the possibility of failures when migrating.